+++ src/types/table.ts
@@ -1,11 +1,10 @@
 export interface Column {
-    label: string;
-    accessor: string;
-    sortable?: boolean;
-    width?: string;
-    render?: (item: Row) => React.ReactNode;
+  label: string;
+  accessor: string;
+  sortable?: boolean;
+  width?: string
 }
 
 export interface Row {
-    [key: string]: string | number | boolean | null;
+  [key: string]: string | number | boolean | null;
 }
